+++ docassemble_webapp/docassemble/webapp/mail/helpers.py
@@ -1,10 +1,9 @@
 import copy
-from flask import current_app
 from docassemble.webapp.config import daconfig
 
 mail_configs = {}
 
-def get_mail_config():
+def get_mail_config(app):
     the_mail_configs = {}
     default_config = None
     for mail_config in daconfig['mail']:
@@ -14,13 +13,13 @@ def get_mail_config():
         default_config = daconfig['mail'][0]
     if default_config is None:
         default_config = {'username': None, 'password': None, 'default sender': None, 'server': 'localhost', 'port': 25, 'use ssl': False, 'use tls': True}
-    current_app.config['MAIL_USERNAME'] = default_config.get('username', None)
-    current_app.config['MAIL_PASSWORD'] = default_config.get('password', None)
-    current_app.config['MAIL_DEFAULT_SENDER'] = default_config.get('default sender', None)
-    current_app.config['MAIL_SERVER'] = default_config.get('server', 'localhost')
-    current_app.config['MAIL_PORT'] = default_config.get('port', 25)
-    current_app.config['MAIL_USE_SSL'] = default_config.get('use ssl', False)
-    current_app.config['MAIL_USE_TLS'] = default_config.get('use tls', True)
+    app.config['MAIL_USERNAME'] = default_config.get('username', None)
+    app.config['MAIL_PASSWORD'] = default_config.get('password', None)
+    app.config['MAIL_DEFAULT_SENDER'] = default_config.get('default sender', None)
+    app.config['MAIL_SERVER'] = default_config.get('server', 'localhost')
+    app.config['MAIL_PORT'] = default_config.get('port', 25)
+    app.config['MAIL_USE_SSL'] = default_config.get('use ssl', False)
+    app.config['MAIL_USE_TLS'] = default_config.get('use tls', True)
     count = 0
     for mail_config in daconfig['mail']:
         the_config = copy.deepcopy(mail_config)
@@ -45,10 +44,10 @@ def get_mail_config():
                 'MAILGUN_API_URL': mail_config['mailgun api url'],
                 'MAILGUN_API_KEY': mail_config['mailgun api key'],
                 'MAIL_DEFAULT_SENDER': mail_config.get('default sender', None),
-                'MAIL_DEBUG': current_app.config.get('MAIL_DEBUG', False),
-                'MAIL_MAX_EMAILS': current_app.config.get('MAIL_MAX_EMAILS'),
-                'MAIL_SUPPRESS_SEND': current_app.config.get('MAIL_SUPPRESS_SEND', False),
-                'MAIL_ASCII_ATTACHMENTS': current_app.config.get('MAIL_ASCII_ATTACHMENTS', False)
+                'MAIL_DEBUG': app.config.get('MAIL_DEBUG', False),
+                'MAIL_MAX_EMAILS': app.config.get('MAIL_MAX_EMAILS'),
+                'MAIL_SUPPRESS_SEND': app.config.get('MAIL_SUPPRESS_SEND', False),
+                'MAIL_ASCII_ATTACHMENTS': app.config.get('MAIL_ASCII_ATTACHMENTS', False)
             }
         elif 'sendgrid api key' in mail_config and mail_config['sendgrid api key']:
             from .sendgrid_mail import Mail as SendgridMail  # pylint: disable=import-outside-toplevel
@@ -56,10 +55,10 @@ def get_mail_config():
             config = {
                 'SENDGRID_API_KEY': mail_config['sendgrid api key'],
                 'MAIL_DEFAULT_SENDER': mail_config.get('default sender', None),
-                'MAIL_DEBUG': current_app.config.get('MAIL_DEBUG', False),
-                'MAIL_MAX_EMAILS': current_app.config.get('MAIL_MAX_EMAILS'),
-                'MAIL_SUPPRESS_SEND': current_app.config.get('MAIL_SUPPRESS_SEND', False),
-                'MAIL_ASCII_ATTACHMENTS': current_app.config.get('MAIL_ASCII_ATTACHMENTS', False)
+                'MAIL_DEBUG': app.config.get('MAIL_DEBUG', False),
+                'MAIL_MAX_EMAILS': app.config.get('MAIL_MAX_EMAILS'),
+                'MAIL_SUPPRESS_SEND': app.config.get('MAIL_SUPPRESS_SEND', False),
+                'MAIL_ASCII_ATTACHMENTS': app.config.get('MAIL_ASCII_ATTACHMENTS', False)
             }
         elif 'azure client id' in mail_config and mail_config['azure client id']:
             from .azure_mail import Mail as AzureMail  # pylint: disable=import-outside-toplevel
@@ -69,10 +68,10 @@ def get_mail_config():
                 'AZURE_CLIENT_SECRET': mail_config.get('azure client secret', None),
                 'AZURE_TENANT_ID': mail_config.get('azure tenant id', None),
                 'MAIL_DEFAULT_SENDER': mail_config.get('default sender', None),
-                'MAIL_DEBUG': current_app.config.get('MAIL_DEBUG', False),
-                'MAIL_MAX_EMAILS': current_app.config.get('MAIL_MAX_EMAILS'),
-                'MAIL_SUPPRESS_SEND': current_app.config.get('MAIL_SUPPRESS_SEND', False),
-                'MAIL_ASCII_ATTACHMENTS': current_app.config.get('MAIL_ASCII_ATTACHMENTS', False)
+                'MAIL_DEBUG': app.config.get('MAIL_DEBUG', False),
+                'MAIL_MAX_EMAILS': app.config.get('MAIL_MAX_EMAILS'),
+                'MAIL_SUPPRESS_SEND': app.config.get('MAIL_SUPPRESS_SEND', False),
+                'MAIL_ASCII_ATTACHMENTS': app.config.get('MAIL_ASCII_ATTACHMENTS', False)
             }
         else:
             from .da_flask_mail import FlaskMail  # pylint: disable=import-outside-toplevel
@@ -85,18 +84,18 @@ def get_mail_config():
                 'MAIL_USE_TLS': mail_config.get('use tls', True),
                 'MAIL_USE_SSL': mail_config.get('use ssl', False),
                 'MAIL_DEFAULT_SENDER': mail_config.get('default sender', None),
-                'MAIL_DEBUG': current_app.config.get('MAIL_DEBUG', False),
-                'MAIL_MAX_EMAILS': current_app.config.get('MAIL_MAX_EMAILS'),
-                'MAIL_SUPPRESS_SEND': current_app.config.get('MAIL_SUPPRESS_SEND', False),
-                'MAIL_ASCII_ATTACHMENTS': current_app.config.get('MAIL_ASCII_ATTACHMENTS', False)
+                'MAIL_DEBUG': app.config.get('MAIL_DEBUG', False),
+                'MAIL_MAX_EMAILS': app.config.get('MAIL_MAX_EMAILS'),
+                'MAIL_SUPPRESS_SEND': app.config.get('MAIL_SUPPRESS_SEND', False),
+                'MAIL_ASCII_ATTACHMENTS': app.config.get('MAIL_ASCII_ATTACHMENTS', False)
             }
         if config_name == 'default':
-            mail_config['mail'] = mail_class(app=current_app, config=config)
+            mail_config['mail'] = mail_class(app=app, config=config)
         else:
             mail_config['mail'] = mail_class(config=config)
     return the_mail_configs
 
 
-def compute_mail_config():
+def compute_mail_config(app):
     mail_configs.clear()
-    mail_configs.update(get_mail_config())
+    mail_configs.update(get_mail_config(app))